+++ include/fft.h
@@ -6,7 +6,7 @@
 #define BLK_SIZE 1024
 
 extern float *fft(size_t count, float *samples,
-				  double *time_upload, double *time_exec, double *time_download,
+				  float *time_upload, float *time_exec, float *time_download,
 				  int profiling);
 
 #endif
