+++ include/mdct.h
@@ -2,6 +2,6 @@
 #define AACL_MDCT_H
 
 extern float *mdct(size_t count, float *samples,
-				   double *time_upload, double *time_exec, double *time_download,
+				   float *time_upload, float *time_exec, float *time_download,
 				   int profiling);
 #endif
